+++ controllers/AdminController.php
@@ -9,6 +9,7 @@
 namespace humhub\modules\externalWebsites\controllers;
 
 use Yii;
+use humhub\modules\admin\components\Controller;
 use humhub\modules\admin\permissions\ManageSettings;
 use humhub\modules\externalWebsites\models\Page;
 use humhub\modules\comment\models\Comment;
@@ -17,7 +18,7 @@ use humhub\modules\comment\models\Comment;
  * AdminController
  * For administrators only
  */
-class AdminController extends \humhub\modules\admin\components\Controller
+class AdminController extends Controller
 {
     /**
      * @inheritdoc